requirements.txt
streamlit>=1.29.0
python-dotenv>=1.0.0
langchain>=0.1.115
langchain-text-splitters
langchain-openai
langchain-community
PyPDF2>=3.0.1
faiss-cpu>=1.13.2
openai>=0.27.10